Python code to update the graph
This code file creates a graph for the streets considering each street as a node, their properties as the node attributes, and the calculated weight (Time*severityProportion) as the attribute of the node relationship. Further, through calculations from traversal, unnecessary relationships have been discarded, as the logic utilized in Predict.py calculated all possible combinations between streets, which is entirely redundant. Once the required edges are calculated, the updated graph is converted to a CSV file and uploaded on Neo4j for traversal calculations. There, the weights, being so small, have been scaled by a power of 1e6 in order to add impact to the calculation.